Move `DateFormatter`s to extension.
